Integrate truthy rewriting into ExprRewriter
