+++ src/transform/expr.rs
@@ -67,6 +67,20 @@ impl<'a> ExprRewriter<'a> {
         self.scopes.borrow_mut().last_mut().unwrap().push(func);
     }
 
+    fn wrap_truthy_expr(&self, expr: &mut Expr) {
+        if !self.options.truthy || is_truth_call(expr) {
+            return;
+        }
+
+        let original = expr.clone();
+        *expr = py_expr!(
+            "
+__dp__.truth({expr:expr})
+",
+            expr = original,
+        );
+    }
+
     fn rewrite_target(&self, target: Expr, value: Expr, out: &mut Vec<Stmt>) {
         match target {
             Expr::Tuple(tuple) => {
@@ -649,12 +663,52 @@ def {func:id}():
         };
 
         walk_stmt(self, stmt);
+
+        if self.options.truthy {
+            match stmt {
+                Stmt::If(ast::StmtIf {
+                    test,
+                    elif_else_clauses,
+                    ..
+                }) => {
+                    self.wrap_truthy_expr(test);
+                    for clause in elif_else_clauses {
+                        if let Some(test) = &mut clause.test {
+                            self.wrap_truthy_expr(test);
+                        }
+                    }
+                }
+                Stmt::While(ast::StmtWhile { test, .. }) => {
+                    self.wrap_truthy_expr(test);
+                }
+                _ => {}
+            }
+        }
+    }
+}
+
+fn is_truth_call(expr: &Expr) -> bool {
+    match expr {
+        Expr::Call(ast::ExprCall {
+            func, arguments, ..
+        }) if arguments.args.len() == 1 && arguments.keywords.is_empty() => match func.as_ref() {
+            Expr::Attribute(ast::ExprAttribute { value, attr, .. }) if attr.as_str() == "truth" => {
+                matches!(
+                    value.as_ref(),
+                    Expr::Name(ast::ExprName { id, .. })
+                        if id.as_str() == "__dp__"
+                )
+            }
+            _ => false,
+        },
+        _ => false,
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::test_util::assert_transform_eq_truthy;
     use crate::transform::Options;
     use ruff_python_codegen::{Generator, Stylist};
     use ruff_python_parser::parse_module;
@@ -1209,5 +1263,56 @@ r = [a * b for a in items for b in items2]
         let output = rewrite_source(input);
         assert!(output.contains("getattr(__dp__, \"iter\")(items)"));
         assert!(output.contains("getattr(__dp__, \"mul\")(a, b)"));
+    }
+
+    #[test]
+    fn rewrites_truthy_if_condition() {
+        let input = r#"
+if a:
+    pass
+else:
+    pass
+"#;
+        let expected = r#"
+if __dp__.truth(a):
+    pass
+else:
+    pass
+"#;
+        assert_transform_eq_truthy(input, expected);
+    }
+
+    #[test]
+    fn rewrites_truthy_elif_and_else() {
+        let input = r#"
+if a:
+    pass
+elif b:
+    pass
+else:
+    pass
+"#;
+        let expected = r#"
+if __dp__.truth(a):
+    pass
+elif __dp__.truth(b):
+    pass
+else:
+    pass
+"#;
+        assert_transform_eq_truthy(input, expected);
+    }
+
+    #[test]
+    fn rewrites_truthy_while_condition() {
+        let input = r#"
+while a:
+    pass
+"#;
+        let expected = r#"
+while __dp__.truth(a):
+    pass
+"#;
+        assert_transform_eq_truthy(input, expected);
     }
 }
